@@ -14,12 +14,13 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@testing-library/react": "^16.1.0",
     "@types/axios": "^0.14.0",
     "@types/faker": "^4.1.11",
     "@types/jest": "^27.5.2",
     "@types/node": "^18.0.1",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
     "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^5.30.5",
     "clean-webpack-plugin": "^4.0.0",
@@ -36,7 +37,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^27.2.0",
     "lint-staged": "^11.1.2",
-    "sass-loader": "^12.1.0",
+    "sass-loader": "^16.0.4",
     "style-loader": "^3.2.1",
     "ts-jest": "^27.0.5",
     "ts-loader": "^9.2.5",
@@ -49,10 +50,10 @@
   "dependencies": {
     "axios": "^0.21.4",
     "http-proxy-middleware": "^2.0.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.28.1",
     "react-scripts": "^4.0.3",
-    "sass": "^1.83.0"
+    "sass": "^1.83.1"
   }
 }
