@@ -1,5 +1,8 @@
 import Router from "@/presentation/components/router/router";
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client"; // Modificado para importar de 'react-dom/client'
+import "@/presentation/styles/global.scss";
 
-ReactDOM.render(<Router />, document.getElementById("main"));
+// Criação do root usando createRoot
+const root = ReactDOM.createRoot(document.getElementById("main"));
+root.render(<Router />);
